Bj capacity 3

+++ src/Interfaces/IFacility.cs
@@ -5,7 +5,7 @@ namespace Trestlebridge.Interfaces
 {
     public interface IFacility<T>
     {
-        double Capacity { get; }
+        int Capacity { get; }
 
         void AddResource (T resource);
         void AddResource (List<T> resources);
